+++ subliminal/providers/podnapisi.py
@@ -18,7 +18,7 @@ from zipfile import ZipFile
 from . import Provider, get_version
 from .. import __version__
 from ..exceptions import ProviderError
-from ..subtitle import Subtitle, fix_line_ending, guess_matches
+from ..subtitle import Subtitle, fix_line_ending, guess_matches, sanitized_string_equal
 from ..video import Episode, Movie
 
 logger = logging.getLogger(__name__)
@@ -47,7 +47,7 @@ class PodnapisiSubtitle(Subtitle):
         # episode
         if isinstance(video, Episode):
             # series
-            if video.series and self.title.lower() == video.series.lower():
+            if video.series and sanitized_string_equal(self.title, video.series):
                 matches.add('series')
             # season
             if video.season and self.season == video.season:
@@ -61,7 +61,7 @@ class PodnapisiSubtitle(Subtitle):
         # movie
         elif isinstance(video, Movie):
             # title
-            if video.title and self.title.lower() == video.title.lower():
+            if video.title and sanitized_string_equal(self.title, video.title):
                 matches.add('title')
             # guess
             for release in self.releases: